pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Viktoriya Dzhoruhova <viktodzho@gmail.com>"]
readme = "README.md"

packages = [
    { include = "gendiff" }
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
gendiff = "gendiff.scripts:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
